The Kurdish flag
Hello, 
Kurdistan is not an official country, it's the land of Kurds, the largest ethnic group in the world without their own state. 
This flag is a an official flag of Kurdistan region (North Iraq). Kurdistan region has officially it's own domain extension .krd. So I have chosen the file name to be krd.svg
Here is more info about the gTLD https://dot.krd/
This svg file is valid, I have checked that on the w3c validator. 

Thanks, 

Peshmerge