Choose flags to include (#402)
Give an option to allow users to specify which flags to include in the output CSS.

Specify using a variable list in SCSS like so:

`$include: 'as', 'bl', 'si';`

If this variable is omitted then all flags are imported as normal.

Also moved the "more" flags into the main list as it seems pointless separating them. Maybe this could also be expanded to have an $exclude variable too so it includes them all apart from the ones specified. It then would allow full control in three different ways!

- All flags
- All but some (exclude)
- Some but not all (include)

Thanks,
Neil