Update Travis to run on latest stable and LTS Node.js
With this configuration, Travis will always run a build on both the
latest stable and LTS Node.js versions.

This is also required to fix the block from Dependabot, Husky 7.0.0
requires Node.js 12 or higher and Travis runs Node.js 10, so CI is
failing on the PR to update Husky.